backend_ai_conversation/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
google-genai>=0.8.0
elevenlabs>=1.2.0
pydantic>=2.0.0
redis>=5.0.0          # optional — only needed if using Redis sessions
